# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Sales Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Prospect {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  companySize?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  status: string;
  /** @nullable */
  campaignId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProspectInput {
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  industry?: string;
  companySize?: string;
  linkedinUrl?: string;
  notes?: string;
  campaignId?: number;
}

export interface ProspectUpdate {
  firstName?: string;
  lastName?: string;
  email?: string;
  company?: string;
  jobTitle?: string;
  industry?: string;
  companySize?: string;
  linkedinUrl?: string;
  notes?: string;
  status?: string;
  /** @nullable */
  campaignId?: number | null;
}

export interface Campaign {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: string;
  productName: string;
  productDescription: string;
  /** @nullable */
  targetIndustry?: string | null;
  senderName: string;
  senderEmail: string;
  tone: string;
  createdAt: string;
  updatedAt: string;
}

export type CampaignInputTone = typeof CampaignInputTone[keyof typeof CampaignInputTone];


export const CampaignInputTone = {
  professional: 'professional',
  friendly: 'friendly',
  casual: 'casual',
  urgent: 'urgent',
} as const;

export interface CampaignInput {
  name: string;
  description?: string;
  productName: string;
  productDescription: string;
  targetIndustry?: string;
  senderName: string;
  senderEmail: string;
  tone: CampaignInputTone;
}

export interface CampaignUpdate {
  name?: string;
  description?: string;
  status?: string;
  productName?: string;
  productDescription?: string;
  targetIndustry?: string;
  senderName?: string;
  senderEmail?: string;
  tone?: string;
}

export interface Email {
  id: number;
  prospectId: number;
  /** @nullable */
  campaignId?: number | null;
  subject: string;
  body: string;
  status: string;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface EmailInput {
  prospectId: number;
  campaignId?: number;
  subject: string;
  body: string;
}

export interface EmailUpdate {
  subject?: string;
  body?: string;
  status?: string;
}

export interface GenerateEmailInput {
  prospectId: number;
  /** @nullable */
  campaignId?: number | null;
  customInstructions?: string;
}

export interface GeneratedEmail {
  subject: string;
  body: string;
}

export interface ActivityItem {
  type: string;
  description: string;
  timestamp: string;
}

export interface DashboardStats {
  totalProspects: number;
  totalCampaigns: number;
  emailsSent: number;
  emailsDraft: number;
  prospectsContacted: number;
  prospectsNew: number;
  recentActivity: ActivityItem[];
}

export interface ApiError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListProspectsParams = {
/**
 * @nullable
 */
campaignId?: number | null;
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
search?: string | null;
};

export type ListEmailsParams = {
/**
 * @nullable
 */
prospectId?: number | null;
/**
 * @nullable
 */
campaignId?: number | null;
/**
 * @nullable
 */
status?: string | null;
};