# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Sales Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /emails:
    get:
      operationId: listEmails
      tags: [emails]
      summary: List all emails
      parameters:
        - name: prospectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"
    post:
      operationId: createEmail
      tags: [emails]
      summary: Save a drafted email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "201":
          description: Created email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get an email by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateEmail
      tags: [emails]
      summary: Update an email (edit draft or change status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailUpdate"
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteEmail
      tags: [emails]
      summary: Delete an email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /emails/{id}/send:
    post:
      operationId: sendEmail
      tags: [emails]
      summary: Mark an email as sent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email marked as sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /emails/generate:
    post:
      operationId: generateEmail
      tags: [emails]
      summary: AI-generate a personalized email draft for a prospect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateEmailInput"
      responses:
        "200":
          description: Generated email content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedEmail"
...
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        productName:
          type: string
        productDescription:
          type: string
        targetIndustry:
          type: ["string", "null"]
        senderName:
          type: string
        senderEmail:
          type: string
        tone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - productName
        - productDescription
        - senderName
        - senderEmail
        - tone
        - createdAt
        - updatedAt
    CampaignInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        productName:
          type: string
        productDescription:
          type: string
        targetIndustry:
          type: string
        senderName:
          type: string
        senderEmail:
          type: string
        tone:
          type: string
          enum: [professional, friendly, casual, urgent]
      required:
        - name
        - productName
        - productDescription
        - senderName
        - senderEmail
        - tone
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        productName:
          type: string
        productDescription:
          type: string
        targetIndustry:
          type: string
        senderName:
          type: string
        senderEmail:
          type: string
        tone:
          type: string
    Email:
      type: object
      properties:
        id:
          type: integer
        prospectId:
          type: integer
        campaignId:
          type: ["integer", "null"]
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        sentAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - prospectId
        - subject
        - body
        - status
        - createdAt
        - updatedAt
    EmailInput:
      type: object
      properties:
        prospectId:
          type: integer
        campaignId:
          type: integer
        subject:
          type: string
        body:
          type: string
      required:
        - prospectId
        - subject
        - body
    EmailUpdate:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        status:
          type: string
    GenerateEmailInput:
      type: object
      properties:
        prospectId:
          type: integer
        campaignId:
          type: ["integer", "null"]
        customInstructions:
          type: string
      required:
        - prospectId
    GeneratedEmail:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
      required:
        - subject
        - body
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...