# lib/api-zod/src/generated/types/campaignInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Sales Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignInputTone } from './campaignInputTone';

export interface CampaignInput {
  name: string;
  description?: string;
  productName: string;
  productDescription: string;
  targetIndustry?: string;
  senderName: string;
  senderEmail: string;
  tone: CampaignInputTone;
}