# lib/api-zod/src/generated/types/email.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Sales Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Email {
  id: number;
  prospectId: number;
  /** @nullable */
  campaignId?: number | null;
  subject: string;
  body: string;
  status: string;
  /** @nullable */
  sentAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}