# lib/api-zod/src/generated/types/generateEmailInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Sales Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateEmailInput {
  prospectId: number;
  /** @nullable */
  campaignId?: number | null;
  customInstructions?: string;
}